feat: detect floating promises
The `@typescript-eslint` plugin has an incredibly useful feature that
will detect floating promises - that is, promises that are not `await`ed
on, returned or otherwise don't have their errors handled.

E.g. it will detect this:

```js
async function foo () {

}

function bar () {
  foo()  // will cause linting to fail
}
```

Possibly contentious, but to enable this, you must be able to derive types
from your code, so all linted code must also be type checked.

We've got a few repos that do not have types for their tests which will need
backfilling before they can use this feature.

As implemented the `@typescript-eslint/no-floating-promises` rule is
enabled when a `tsconfig.json` is detected in a project - this will mean
projects that skip typing their tests will break if this is merged.

Harsh, perhaps, but detecting this class of error is incredibly useful.