fix: allow `any` args and referencing functions by property name
Disables two rules:

`@typescript-eslint/no-unsafe-argument` - this prevents passing vars
with an `any` type to functions.

Unfortunately we still deal with untyped modules so this will happen.

`@typescript-eslint/unbound-method` - this prevents referencing functions
without invoking them.

We need to disable this to do things like pass functions passed as
part of an options object on to other functions.